% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jfi.R
\name{jfi_plot}
\alias{jfi_plot}
\title{Create a ggplot with JFI Theme and Default Caption}
\usage{
jfi_plot(...)
}
\arguments{
\item{...}{Arguments passed to `ggplot()`, such as `data` and `aes()`.}
}
\value{
A `ggplot` object with the JFI theme and default or custom caption.
}
\description{
`jfi_plot` wraps around `ggplot()` to automatically apply the JFI custom theme and insert a default caption if none is provided.
}
\examples{
jfi_plot(mtcars, aes(x = wt, y = mpg)) + geom_point()

jfi_plot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  labs(caption = caption_w_logo(source = "Data Source"))
}
