% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jfi.R
\name{scale_color_jfi}
\alias{scale_color_jfi}
\title{JFI Color Scale}
\usage{
scale_color_jfi(
  type = c("discrete", "divergent", "sequential", "sequential_discrete"),
  palette = c("jfi", "pw_blue", "pw_red"),
  midpoint = NULL,
  discrete_values = NULL
)
}
\arguments{
\item{type}{The type of scale: 'discrete', 'divergent', 'sequential', 'sequential_discrete'.}

\item{palette}{The color palette to use: 'jfi', 'pw_blue', 'pw_red'.}

\item{midpoint}{The midpoint for the divergent scale (if applicable).}

\item{discrete_values}{The discrete values for the sequential discrete scale (if applicable).}
}
\value{
A ggplot2 scale object.
}
\description{
Custom color scale based on the JFI style guide.
}
\examples{
ggplot(mtcars, aes(x = factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_jfi(type = "discrete", palette = "jfi")
}
