% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jfi.R
\name{set_jfi_defaults}
\alias{set_jfi_defaults}
\title{Set JFI ggplot2 Defaults}
\usage{
set_jfi_defaults(
  base_size = 8,
  base_family = "melior",
  base_line_size = base_size/170,
  base_rect_size = base_size/170,
  scale = "continuous"
)
}
\arguments{
\item{base_size}{The base font size for the theme. All fonts are relative to this value.}

\item{base_family}{The base font family for the theme.}

\item{base_line_size}{The base line size for the theme. All line sizes are relative to this value.}

\item{base_rect_size}{The base rect size for the theme. All rect sizes are relative to this value.}
}
\description{
\code{set_jfi_defaults} sets the default ggplot2 theme, font, and geom aesthetics for the R session.
}
