% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jfi.R
\name{theme_jfi}
\alias{theme_jfi}
\title{Custom JFI Theme}
\usage{
theme_jfi(
  legend_right = FALSE,
  base_size = 8,
  base_family = "melior",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{legend_right}{Logical, should the legend be on the right?}

\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base line size.}

\item{base_rect_size}{Base rectangle size.}
}
\value{
A ggplot2 theme object.
}
\description{
A custom ggplot2 theme based on the JFI style guide.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_jfi()

}
